Move Zoom setting and UIActions to RichTextBoxBase so SettingsTextBox automatically receives the same actions, and initializes to the auto-saved zoom factor.
