+++ Sandra.UI.WF.Chess/MovesTextBox.cs
@@ -72,12 +72,6 @@ namespace Sandra.UI.WF
 
             applyDefaultStyle();
 
-            int zoomFactor;
-            if (Program.TryGetAutoSaveValue(SettingKeys.Zoom, out zoomFactor))
-            {
-                ZoomFactor = PType.RichTextZoomFactor.FromDiscreteZoomFactor(zoomFactor);
-            }
-
             // DisplayTextChanged handlers are called immediately upon registration.
             // This initializes moveFormatter.
             localizedPieceSymbols.DisplayText.ValueChanged += _ =>
@@ -452,17 +446,6 @@ namespace Sandra.UI.WF
             }
 
             base.OnMouseWheel(e);
-
-            if (ModifierKeys.HasFlag(Keys.Control))
-            {
-                // ZoomFactor isn't updated yet, so predict here what it's going to be.
-                autoSaveZoomFactor(PType.RichTextZoomFactor.ToDiscreteZoomFactor(ZoomFactor) + Math.Sign(e.Delta));
-            }
-        }
-
-        private void autoSaveZoomFactor(int zoomFactor)
-        {
-            Program.AutoSave.Persist(SettingKeys.Zoom, zoomFactor);
         }
     }
 }
