+++ Sandra.UI.WF.Chess/RichTextBoxBase.UIActions.cs
@@ -57,5 +57,31 @@ namespace Sandra.UI.WF
             if (perform) SelectAll();
             return UIActionVisibility.Enabled;
         }
+
+        public UIActionState TryZoomIn(bool perform)
+        {
+            int zoomFactor = PType.RichTextZoomFactor.ToDiscreteZoomFactor(ZoomFactor);
+            if (zoomFactor >= PType.RichTextZoomFactor.MaxDiscreteValue) return UIActionVisibility.Disabled;
+            if (perform)
+            {
+                zoomFactor++;
+                ZoomFactor = PType.RichTextZoomFactor.FromDiscreteZoomFactor(zoomFactor);
+                autoSaveZoomFactor(zoomFactor);
+            }
+            return UIActionVisibility.Enabled;
+        }
+
+        public UIActionState TryZoomOut(bool perform)
+        {
+            int zoomFactor = PType.RichTextZoomFactor.ToDiscreteZoomFactor(ZoomFactor);
+            if (zoomFactor <= PType.RichTextZoomFactor.MinDiscreteValue) return UIActionVisibility.Disabled;
+            if (perform)
+            {
+                zoomFactor--;
+                ZoomFactor = PType.RichTextZoomFactor.FromDiscreteZoomFactor(zoomFactor);
+                autoSaveZoomFactor(zoomFactor);
+            }
+            return UIActionVisibility.Enabled;
+        }
     }
 }
